fix: build error, clean code, tweaks
