+++ packages/unconfig-vitepress/src/plugins/nolebase/index.ts
@@ -98,7 +98,6 @@ const defaultOptions: NolebasePluginPresetOptions = {
   gitChangelog: {
     enable: true,
     options: {
-      mapContributors: [],
       commitsRelativeTime: true,
     },
   },
